+++ epyk/core/Page.py
@@ -39,22 +39,21 @@ class JsProperties:
 
     @property
     def frgs(self):
-        """ Return the extra JavaScript function manually added. """
+        """Return the extra JavaScript function manually added. """
         return self._context["text"]
 
     @property
     def constructors(self):
-        """ Get the list of needed contractors """
+        """Get the list of needed contractors """
         return self._context['constructors'].keys()
 
     @property
     def functions(self):
-        """ Get the list of needed javasScript functions """
+        """Get the list of needed javasScript functions """
         return self._context['functions'].keys()
 
     def add_text(self, text: str, map_id: str = None):
-        """
-        Add JavaScript fragments from String.
+        """Add JavaScript fragments from String.
 
         Usage::
 
@@ -71,8 +70,7 @@ class JsProperties:
                 self.__map_css.add(map_id)
 
     def add_event(self, event: str, value):
-        """
-        Add JavaScript fragments from String.
+        """Add JavaScript fragments from String.
 
         :param event: JavaScript fragments to be directly included to the page
         :param value:
@@ -80,8 +78,7 @@ class JsProperties:
         self._context["events"][event] = value
 
     def add_builders(self, builder_def: str, func_dsc: str = None):
-        """
-        This will use add or extend according to the builder_def type.
+        """This will use add or extend according to the builder_def type.
 
         #TODO implement func_dsc
 
@@ -98,8 +95,7 @@ class JsProperties:
                 self._context['builders'].append(builder_def)
 
     def add_on_ready(self, builder_def: str):
-        """
-        Add JavaScript expression in the onReady selection of the HTML page.
+        """Add JavaScript expression in the onReady selection of the HTML page.
 
         Usage::
 
@@ -112,8 +108,7 @@ class JsProperties:
         self._context['onReady'].add(builder_def)
 
     def add_constructor(self, name: str, content: str, override: bool = False, verbose: bool = False) -> str:
-        """
-        Register the constructor function and return its reference.
+        """Register the constructor function and return its reference.
 
         :param name: The constructor name
         :param content: The entire definition
@@ -128,21 +123,19 @@ class JsProperties:
         self._context['constructors'][name] = content
         return name
 
-    def has_constructor(self, name: str):
-        """
-        Check if a constructor is already defined.
+    def has_constructor(self, name: str) -> bool:
+        """Check if a constructor is already defined.
 
         :param name: The constructor name
         """
         return name in self._context['constructors']
 
     def get_constructor(self, name: str):
-        """ Get the constructor definition. """
+        """Get the constructor definition. """
         return self._context['constructors'].get(name)
 
     def set_constructor(self, name: str, content: str = None, func_ref: bool = False):
-        """
-        Set a constructor.
+        """Set a constructor.
 
         Usage::
 
@@ -151,7 +144,7 @@ class JsProperties:
           btn = page.ui.button("click")
           btn.click([btn.build("Clicked")])
 
-        :param name: The constructor alias (must of the same name than the function)
+        :param name: The constructor alias (must match the function's name)
         :param content: The constructor expression in JavaScript
         :param func_ref:
         """
@@ -163,9 +156,7 @@ class JsProperties:
             self._context['constructors'][name] = content
 
     def add_function(self, name: str, js_funcs: Union[list, str], pmts: list) -> dict:
-        """
-        Add a JavaScript function to the page.
-
+        """Add a JavaScript function to the page.
         This method will build the JavaScript method and it will be in charge during the conversion to JavaScript to
         write function [name] ( [pmts] ){ [js_funcs] }
 
@@ -185,18 +176,18 @@ class JsProperties:
         return self._context['functions'][name]
 
     def get_function(self, name: str) -> str:
-        """ Get an existing function definition. """
+        """Get an existing function definition. """
         if name in self._context['functions']:
             func = self._context['functions'][name]
             return "function %s(%s){%s}" % (name, ", ".join(func['pmt']), func['content'])
 
     def set_init_options(self, html_code: str, options: str):
-        """ Set init options to the main scope of the report """
+        """Set init options to the main scope of the report """
         if isinstance(html_code, str):
             self._context['init'][html_code] = options
 
     def get_init_options(self, html_code: str = None) -> dict:
-        """ get init option of a component """
+        """Get init option of a component """
         if html_code is None:
             return self._context['init']
 
@@ -211,12 +202,11 @@ class CssProperties:
 
     @property
     def text(self) -> str:
-        """ Return the extra CSS styles manually added. """
+        """Return the extra CSS styles manually added. """
         return "\n".join(self._context['css']["text"])
 
     def add_text(self, text: str, map_id: str = None, replace: bool = False):
-        """
-        Add CSS style from String.
+        """Add CSS styles from String.
 
         Usage::
 
@@ -238,13 +228,12 @@ class CssProperties:
             self._context['css']["text"][self.__map_css[map_id]] = text
 
     def remove_text(self, map_id: str):
-        """
-        Remove th text from the mapping.
+        """Remove th text from the mapping.
         This will also reset the following indices.
 
-        :param map_id: Optional. Internal ID to avoid loading the same content multiple time
+        :param map_id: Internal ID to avoid loading the same content multiple time
         """
-        if  map_id  in self.__map_css:
+        if map_id in self.__map_css:
             i = self.__map_css[map_id]
             self._context['css']["text"].pop(i)
             tmp_css_context = {}
@@ -258,8 +247,7 @@ class CssProperties:
             self.__map_css = tmp_css_context
 
     def add_builders(self, builder_def: str):
-        """
-        This will use add or extend to the CSS JavaScript builders according to the builder_def type.
+        """This will use add or extend to the CSS JavaScript builders according to the builder_def type.
 
         :param builder_def: The builder definition
         """
@@ -271,16 +259,14 @@ class CssProperties:
             self._context['js']['builders_css'].append(builder_def)
 
     def container_style(self, css: dict):
-        """
-        Set the container CSS style.
+        """Set the container CSS style.
 
         :param css: The CSS attributes to be applied
         """
         self._context['css']['container'].update(css)
 
     def font_face(self, font_family: str, src, stretch: str = "normal", style: str = "normal", weight: str = "normal"):
-        """
-        Set the font.
+        """Set the font.
 
         :param font_family: Defines the name of the font
         :param src: Defines the URL(s) where the font should be downloaded from
@@ -300,19 +286,17 @@ class Properties:
 
     @property
     def context(self):
-        """ Return the common Page context. """
+        """Return the common Page context. """
         return self._context['context']
 
     @property
     def icon(self):
-        """ Return the page icons definition. """
+        """Return the page icons definition. """
         return self._context['icon']
 
     @property
     def js(self) -> JsProperties:
-        """
-        The JavaScript page properties.
-
+        """The JavaScript page properties.
         This will keep track of all the global functions used by the components.
         """
         if self.__js is None:
@@ -321,9 +305,7 @@ class Properties:
 
     @property
     def css(self) -> CssProperties:
-        """
-        The Css page properties.
-
+        """The CSS page properties.
         This will keep track of all the global functions used by the components.
         CSS Properties will work on both the CSS and JS section of the underlying prop dictionary.
         """
@@ -333,9 +315,7 @@ class Properties:
 
     @property
     def data(self) -> data.DataProperties:
-        """
-        Get the predefined data structures.
-
+        """Get the predefined data structures.
         Those data structures are mainly helper to input the different libraries available in this UI framework.
         """
         if "data" in self._context:
@@ -347,8 +327,7 @@ class Properties:
 
 
 class Report:
-    """
-    Main entry point for any web UI.
+    """Main entry point for any web UI.
 
     This class will store all the HTML components, JavaScript fragments and CSS definition in order to then render
     a rich web page.
@@ -371,10 +350,10 @@ class Report:
         This interface will create a generic ui schema which will be then converted.
         For the moment only few output formats are available and web UI are preferred.
 
-        The target of this components is not to replaced libraries but only to easy the use of the most popular ones without
-        having to change languages, This module will help with the versioning, the interactivity and the documentation.
-
-        Indeed libraries interfaces will point to the real documentation to allow you to learn by using it.
+        The target of those components is not to replaced libraries but only to easy the use of the most popular ones
+        without having to change languages, This module will help with the versioning, the interactivity and the
+        documentation.
+        Indeed, libraries interfaces will point to the real documentation to allow you to learn by using it.
 
         Usage::
 
@@ -440,14 +419,14 @@ class Report:
 
     @property
     def properties(self) -> Properties:
-        """ Property to the different Page properties JavaScript and CSS. """
+        """Property to the different Page properties JavaScript and CSS. """
         if self.__properties is None:
             self.__properties = Properties(self._props)
         return self.__properties
 
     @property
     def root__script(self) -> str:
-        """ Return the name of the script creating the Page object. """
+        """Return the name of the script creating the Page object. """
         return self._props["context"]["script"]
 
     @property
@@ -497,9 +476,7 @@ class Report:
 
     @property
     def skins(self) -> skins.Skins:
-        """
-        Add a special skin to the page.
-
+        """Add a special skin to the page.
         This could be used for special event or season during the year (Christmas for example).
 
         Usage::
@@ -527,18 +504,15 @@ class Report:
 
     @property
     def imports(self) -> Imports.ImportManager:
-        """
-        Return the :doc:`report/import_manager`, which allows to import automatically packages for certain components to
-        run.
-
-        By default the imports are retrieved online from CDNJS paths.
+        """Return the :doc:`report/import_manager`, which allows importing automatically packages for certain
+        components to run.
+        By default, the imports are retrieved online from CDNJS paths.
         This can be changed by loading the packages locally and switching off the online mode.
 
         Usage::
 
           page = Report()
           page.imports.setVersion(page.imports.pkgs.popper_js.alias, "1.00.0")
-
         """
         if self.__import_manage is None:
             self.__import_manage = Imports.ImportManager(page=self)
@@ -546,9 +520,7 @@ class Report:
 
     @property
     def symbols(self) -> symboles.Symboles:
-        """
-        Shortcut to the HTML symbols.
-
+        """Shortcut to the HTML symbols.
         Those can be added in string in order to improve the render of a text.
 
         Usage::
@@ -565,9 +537,7 @@ class Report:
 
     @property
     def entities(self) -> entities.Entities:
-        """
-        Shortcut to the HTML Entities.
-
+        """Shortcut to the HTML Entities.
         Those can be added in string in order to improve the render of a text.
 
         Usage::
@@ -583,9 +553,7 @@ class Report:
 
     @property
     def ui(self) -> Interface.Components:
-        """
-        User Interface section.
-
+        """User Interface section.
         All the :doc:`components <report/ui>` which can be used in the dashboard to display the data.
         Within this object different categories of items can be used like (list, simple text, charts...).
 
@@ -594,17 +562,13 @@ class Report:
           page = Report()
           page.ui.text("This is a text")
 
-        Related Pages:
-
-          https://www.w3schools.com/html/default.asp
+        `w3schools <https://www.w3schools.com/html/default.asp>`_
         """
         return self.web.std
 
     @property
     def web(self) -> Interface.WebComponents:
-        """
-        User Interface section.
-
+        """User Interface section.
         All the :doc:`components <report/ui>` which can be used in the dashboard to display the data.
         Within this object different categories of items can be used like (list, simple text, charts...).
 
@@ -613,9 +577,7 @@ class Report:
           page = Report()
           page.web
 
-        Related Pages:
-
-          https://www.w3schools.com/html/default.asp
+        `w3schools <https://www.w3schools.com/html/default.asp>`_
         """
         if self._ui is None:
             self._ui = Interface.WebComponents(self)
@@ -623,8 +585,7 @@ class Report:
 
     @property
     def css(self) -> Classes.Catalog:
-        """
-        Returns the set of :doc:`CSS Classes <css>` for the HTML report.
+        """Returns the set of :doc:`CSS Classes <css>` for the HTML report.
 
         Usage::
 
@@ -637,9 +598,7 @@ class Report:
 
     @property
     def js(self) -> js.Js.JsBase:
-        """
-        Go to the Javascript section. Property to get all the JavaScript features.
-
+        """Go to the Javascript section. Property to get all the JavaScript features.
         Most of the standard modules will be available in order to add event and interaction to the Js transpiled.
 
         Usage::
@@ -650,9 +609,7 @@ class Report:
           page.js.accounting.add_to_imports()
           page.js.moment.add_to_imports()
 
-        Related Pages:
-
-          https://www.w3schools.com/js/default.asp
+        `w3schools <https://www.w3schools.com/html/default.asp>`_
         """
         if self._js is None:
             self._js = js.Js.JsBase(self)
@@ -660,9 +617,7 @@ class Report:
 
     @property
     def py(self) -> PyExt.PyExt:
-        """
-        Python external module section.
-
+        """Python external module section.
         Those are pre-defined Python function to simplify the use of the various components.
 
         Usage::
@@ -670,9 +625,7 @@ class Report:
           page = Report()
           page.py.dates.today()
 
-        Related Pages:
-
-          https://www.w3schools.com/js/default.asp
+        `w3schools <https://www.w3schools.com/html/default.asp>`_
         """
         if self._py is None:
             self._py = PyExt.PyExt(self)
@@ -680,17 +633,14 @@ class Report:
 
     @property
     def auth(self) -> auth.Auth:
-        """
-        Auth interface to allow easy sign-in pages.
+        """Auth interface to allow easy sign-in pages.
 
         Usage::
 
           page = Report()
           page.auth.
 
-        Related Pages:
-
-          https://developers.google.com/identity/sign-in/web/sign-in
+        `Google SignIn <https://developers.google.com/identity/sign-in/web/sign-in>`_
 
         :return: Python Auth Object.
         """
@@ -700,9 +650,7 @@ class Report:
 
     @property
     def data(self) -> data.Data.DataSrc:
-        """
-        Python internal data source management.
-
+        """Python internal data source management.
         This can be extended by inheriting from this epyk.core.data.DataSrc.DataSrc and adding extra entry points.
 
         Usage::
@@ -714,10 +662,8 @@ class Report:
         return data.Data.DataSrc(self)
 
     def register(self, ext_components: Union[list, dict]):
-        """
-        This function allows you to register external Components (namely coming from Pyk Reports)
+        """This function allows you to register external Components (namely coming from Pyk Reports)
         by registering them you this will engrave the object within your report.
-
         The example below will add obj1 and obj2 from an external pyk report previously required,
         then create a div and then add obj3 from an external file.
 
@@ -759,9 +705,7 @@ class Report:
             self.components[comp.html_code] = comp
 
     def get_components(self, html_codes: Union[list, str]):
-        """
-        Retrieve the components based on their ID.
-
+        """Retrieve the components based on their ID.
         This should be used when the htmlCode is defined for a component.
 
         Usage::
@@ -778,8 +722,7 @@ class Report:
         return [self.components[html_code] for html_code in html_codes]
 
     def get_components_by_ref(self, ref: str) -> list:
-        """
-        Get all components defined in the current context by reference.
+        """Get all components defined in the current context by reference.
         Reference can be associated to multiple components.
 
         :param ref: The reference value
@@ -791,15 +734,12 @@ class Report:
         return results
 
     def framework(self, name: str):
-        """
-        Flag to change the way code is transpiled in order to fit with the destination framework.
-
-        By default the code transpiled will be used from a browser in plain Vanilla Js but this will be extended to then
-        be compatible with other framework in order to simplify the path to production and the collaboration between teams.
-
+        """Flag to change the way code is transpiled in order to fit with the destination framework.
+        By default, the code transpiled will be used from a browser in plain Vanilla Js but this will be extended to
+        then be compatible with other framework in order to simplify the path to production and the collaboration
+        between teams.
         Many framework will be compatible like React, Angular, Vue but also some features will be exposed to Kotlin for
         mobile generation.
-
         This work is still in progress.
 
         Usage::
@@ -814,12 +754,9 @@ class Report:
 
     @property
     def outs(self) -> PyOuts.PyOuts:
-        """
-        Link to the possible output formats for a page.
-
-        This will transpile the Python code to web artifacts. Those outputs are standard outputs files in web development.
-
-        The property framework should be used to link to other web framework.
+        """Link to the possible output formats for a page.
+        This will transpile the Python code to web artifacts. Those outputs are standard outputs files in web
+        development. The property framework should be used to link to other web framework.
 
         Usage::
 
@@ -831,8 +768,7 @@ class Report:
 
     @property
     def headers(self) -> html.Header.Header:
-        """
-        Property to the HTML page header.
+        """Property to the HTML page header.
 
         Usage::
 
@@ -847,13 +783,10 @@ class Report:
         return self._header_obj
 
     def dumps(self, result: dict):
-        """
-        Function used to dump the data before being sent to the Javascript layer.
-
+        """Function used to dump the data before being sent to the Javascript layer.
         This function relies on json.dumps with a special encoder in order to work with Numpy array and Pandas data
         structures.
-
-        As NaN is not valid on the Json side those object are not allowed during the dump.
+        As NaN is not valid on the Json side those objects are not allowed during the dump.
         It is advised to use fillna() in your script before returning the data to the framework to avoid this issue.
 
         Usage::
@@ -869,13 +802,11 @@ class Report:
 
         :return: The serialised data
         """
-        return json.dumps(data, cls=js.JsEncoder.Encoder, allow_nan=False)
+        return json.dumps(result, cls=js.JsEncoder.Encoder, allow_nan=False)
 
     @property
     def apps(self) -> apps.AppRoute:
-        """
-        Change the report to a web application.
-
+        """Change the report to a web application.
         This will add extra features available on the target framework.
         For example this HTML page can be transformed to an Angular app, a React App or a Vue one.
 
@@ -891,8 +822,7 @@ class Report:
 
     @property
     def icons(self) -> Icons.IconModel:
-        """
-        Change the icons framework used in the page.
+        """Change icons' framework used in the page.
         Defaults.py in the CSS module is to change the framework for all the page generated by the framework.
 
         Usage::
@@ -910,8 +840,7 @@ class Report:
         return self.__icons
 
     def onDOMContentLoaded(self, js_funcs: Union[str, list], profile: Optional[Union[dict, bool]] = False):
-        """
-        The DOMContentLoaded event fires when the HTML document has been completely parsed, and all deferred scripts
+        """The DOMContentLoaded event fires when the HTML document has been completely parsed, and all deferred scripts
         (<script defer src="…"> and <script type="module">) have downloaded and executed.
         It doesn't wait for other things like images, subframes, and async scripts to finish loading.
 
@@ -921,9 +850,7 @@ class Report:
           div = page.ui.div("Simple Div")
           page.onDOMContentLoaded([div.dom.css({"border": "1px solid red"})])
 
-        Related Pages:
-
-          https://developer.mozilla.org/fr/docs/Web/API/Window/DOMContentLoaded_event
+        `Mozlla <https://developer.mozilla.org/fr/docs/Web/API/Window/DOMContentLoaded_event>`_
 
         :param js_funcs: The Javascript functions to be added to this section
         :param profile: Optional. A flag to set the component performance storage